requirements.txt
accelerate>=0.21.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
datasets==4.5.0
huggingface_hub==1.3.5
ipykernel==7.1.0
matplotlib==3.10.8
ninja==1.13.0
numpy==1.26.4
open_clip_torch==3.2.0
pandas==3.0.0
pillow==12.1.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers>=4.30.0
